+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 - Fixes AqlParseException while using boolean in where clause ([#338](https://github.com/ehrbase/openEHR_SDK/pull/338))
 - Fixes default ASC value for ORDER BY clause ([#342](https://github.com/ehrbase/openEHR_SDK/pull/342))
 - Fixes null nodeIds and annotations missing (https://github.com/ehrbase/openEHR_SDK/pull/343)
+- Fixes duration validation ([#346](https://github.com/ehrbase/openEHR_SDK/pull/346)
 
 ## 1.17.0